requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==4.3.0
beautifulsoup4==4.13.5
blinker==1.9.0
Brotli==1.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.7
dnspython==2.7.0
dulwich==0.24.1
Flask==3.1.2
flask-unsign==1.2.1
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
inflate64==1.0.3
invoke==2.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multivolumefile==0.2.3
paramiko==4.0.0
prettytable==0.7.2
progressbar2==4.5.0
psutil==7.0.0
py7zr==1.0.0
pybcj==1.0.6
pycparser==2.22
pycryptodome==3.23.0
pycryptodomex==3.23.0
PyNaCl==1.6.0
pyppmd==1.2.0
PySocks==1.7.1
python-utils==3.9.1
pyzipper==0.3.6
pyzstd==0.17.0
rarfile==4.2
requests==2.32.5
requests_pkcs12==1.27
smmap==5.0.2
soupsieve==2.7
texttable==1.7.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
